Fix Update-with-Start poll of completed Update
